breakbeam

+++ src/software/power/pins.h
@@ -19,7 +19,7 @@ const uint8_t CHRG_DONE     = 26;
 const uint8_t CHRG          = 25;
 
 // Break Beam
-const uint8_t BREAK_BEAM_PIN = 37;
+const int BREAK_BEAM_PIN = 37;
 
 // UART
 const uint8_t RXD2 = 19;
